facade implementation of the reporter
