+++ src/reporter/Reporter.ts
@@ -1,428 +1,178 @@
-/* eslint-disable no-console */
-import fs from "fs";
-import debug from "debug";
-import chalk from "chalk";
-import CliTable3 from "cli-table3";
-
-import { emoji } from "hardhat/internal/cli/emoji";
-import { pluralize } from "hardhat/internal/util/strings";
-
 import { VerifierLanguageType } from "@solarity/zkit";
-
-import { SpinnerProcessor } from "./SpinnerProcessor";
-import { ProgressBarProcessor } from "./ProgressBarProcessor";
-import { HardhatZKitError } from "../errors";
-import { BYTES_IN_MB } from "../constants";
 
 import { CircuitArtifact } from "../types/artifacts/circuit-artifacts";
 import { CompilationInfo, CircuitSetupInfo } from "../types/core";
+import {
+  ProgressReporter,
+  CircuitFilesResolvingReporter,
+  SetupReporter,
+  CircomCompilerReporter,
+  CircuitCompilationReporter,
+  PtauFileReporter,
+  VerifiersGenerationReporter,
+  VKeyFilesGenerationReporter,
+  ZKeyFilesGenerationReporter,
+} from "./modules";
 
-class BaseReporter {
-  private _spinnerProcessor: SpinnerProcessor;
-  private _progressBarProcessor: ProgressBarProcessor;
+class ReporterFacade {
+  private _setupReporter!: SetupReporter;
+  private _progressReporter!: ProgressReporter;
+  private _ptauFileReporter!: PtauFileReporter;
+  private _circomCompilerReporter!: CircomCompilerReporter;
+  private _circuitCompilationReporter!: CircuitCompilationReporter;
+  private _verifiersGenerationReporter!: VerifiersGenerationReporter;
+  private _vKeyFilesGenerationReporter!: VKeyFilesGenerationReporter;
+  private _zKeyFilesGenerationReporter!: ZKeyFilesGenerationReporter;
+  private _circuitFilesResolvingReporter!: CircuitFilesResolvingReporter;
 
-  constructor(private _quiet: boolean) {
-    this._spinnerProcessor = new SpinnerProcessor();
-    this._progressBarProcessor = new ProgressBarProcessor();
+  constructor(quiet: boolean) {
+    this._initReporters(quiet);
+  }
+
+  private _initReporters(quiet: boolean) {
+    this._setupReporter = new SetupReporter(quiet);
+    this._progressReporter = new ProgressReporter(quiet);
+    this._ptauFileReporter = new PtauFileReporter(quiet);
+    this._circomCompilerReporter = new CircomCompilerReporter(quiet);
+    this._circuitCompilationReporter = new CircuitCompilationReporter(quiet);
+    this._verifiersGenerationReporter = new VerifiersGenerationReporter(quiet);
+    this._vKeyFilesGenerationReporter = new VKeyFilesGenerationReporter(quiet);
+    this._zKeyFilesGenerationReporter = new ZKeyFilesGenerationReporter(quiet);
+    this._circuitFilesResolvingReporter = new CircuitFilesResolvingReporter(quiet);
   }
 
   public setQuiet(newValue: boolean) {
-    this._quiet = newValue;
-  }
-
-  public reportCircuitListToCompile(filteredSourceNames: string[], filteredSourceNamesToCompile: string[]) {
-    if (this.isQuiet()) return;
-
-    if (filteredSourceNamesToCompile.length > 0) {
-      let filesToCompileMessage: string = `\n${chalk.bold("Circuits to compile:")}\n`;
-
-      for (const sourceName of filteredSourceNamesToCompile) {
-        filesToCompileMessage += `\n${chalk.green(">")} ${chalk.italic(sourceName)}`;
-      }
-
-      console.log(filesToCompileMessage);
-    }
-
-    const skippedSourceNames: string[] = filteredSourceNames.filter(
-      (sourceName: string) => !filteredSourceNamesToCompile.includes(sourceName),
-    );
-
-    if (skippedSourceNames.length > 0) {
-      let skippedFilesMessage: string = `\n${chalk.bold("Compilation skipped for:")}\n`;
-
-      for (const sourceName of skippedSourceNames) {
-        skippedFilesMessage += `\n${chalk.yellow(">")} ${chalk.italic.grey(sourceName)}`;
-      }
-
-      console.log(skippedFilesMessage);
-    }
-  }
-
-  public reportCompilerVersion(compilerVersion: string) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.bold("Compiler version:")} ${chalk.green(compilerVersion)}\n`;
-
-    console.log(output);
+    this._initReporters(newValue);
   }
 
   public reportCircuitFilesResolvingStartWithSpinner(): string | null {
-    return this._startSpinner("circuits", "files-resolving", `Resolving and parsing circuits`);
+    return this._circuitFilesResolvingReporter.reportStartWithSpinner();
   }
 
   public reportCircuitFilesResolvingResult(spinnerId: string | null) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const resolvingTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.succeedSpinner(spinnerId, `Circuits are ready for the compilation ${resolvingTimeMessage}`);
+    this._circuitFilesResolvingReporter.reportResult(spinnerId);
   }
 
   public reportCircuitFilesResolvingFail(spinnerId: string | null) {
-    if (this.isQuiet() || !spinnerId) return;
+    this._circuitFilesResolvingReporter.reportFail(spinnerId);
+  }
 
-    const resolvingTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.failSpinner(spinnerId, `Failed to resolve circuit files ${resolvingTimeMessage}\n`);
+  public reportCompilerVersion(compilerVersion: string) {
+    this._circomCompilerReporter.reportVersion(compilerVersion);
   }
 
   public reportCompilationProcessHeader() {
-    if (this.isQuiet()) return;
+    this._circuitCompilationReporter.reportHeader();
+  }
 
-    let output: string = "";
-
-    output += `\n${chalk.bold("Starting compilation process:")}\n`;
-
-    console.log(output);
+  public reportCircuitListToCompile(filteredSourceNames: string[], filteredSourceNamesToCompile: string[]) {
+    this._circuitCompilationReporter.reportCircuitListToCompile(filteredSourceNames, filteredSourceNamesToCompile);
   }
 
   public reportCircuitCompilationStartWithSpinner(circuitName: string, circuitFileName: string): string | null {
-    const fileNameMessage: string = circuitName === circuitFileName ? "" : chalk.grey(` (${circuitFileName}.circom)`);
-
-    return this._startSpinner(circuitName, "compile", `Compiling ${circuitName}${fileNameMessage} circuit`);
+    return this._circuitCompilationReporter.reportStartWithSpinner(circuitName, circuitFileName);
   }
 
   public reportCircuitCompilationResult(spinnerId: string | null, circuitName: string, circuitFileName: string) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const fileNameMessage: string = circuitName === circuitFileName ? "" : chalk.grey(` (${circuitFileName}.circom)`);
-    const compilationTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.succeedSpinner(
-      spinnerId,
-      `Compiled ${chalk.italic(circuitName)}${fileNameMessage} ${compilationTimeMessage}`,
-    );
+    this._circuitCompilationReporter.reportResult(spinnerId, circuitName, circuitFileName);
   }
 
   public reportCircuitCompilationFail(spinnerId: string | null, circuitName: string, circuitFileName: string) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const fileNameMessage: string = circuitName === circuitFileName ? "" : chalk.grey(` (${circuitFileName}.circom)`);
-    const compilationTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.failSpinner(
-      spinnerId,
-      `Failed to compile ${chalk.italic(circuitName)}${fileNameMessage} ${compilationTimeMessage}\n`,
-    );
+    this._circuitCompilationReporter.reportFail(spinnerId, circuitName, circuitFileName);
   }
 
   public reportCompilationResult(compilationInfoArr: CompilationInfo[]) {
-    if (this.isQuiet()) return;
+    this._circuitCompilationReporter.reportFinalResult(compilationInfoArr);
+  }
 
-    let output: string = "";
-    const circuitsMessage: string =
-      compilationInfoArr.length > 1 ? `${compilationInfoArr.length} circuits` : `one circuit`;
+  public reportPtauFileInfo(maxConstraintsNumber: number, ptauId: number, ptauFileFullPath?: string) {
+    this._ptauFileReporter.reportInfo(maxConstraintsNumber, ptauId, ptauFileFullPath);
+  }
 
-    output += `\n${chalk.bold(`Successfully compiled ${circuitsMessage}.`)}\n`;
+  public reportPtauFileDownloadingInfo(ptauFilePath: string, downloadUrl: string) {
+    this._ptauFileReporter.reportDownloadingInfo(ptauFilePath, downloadUrl);
+  }
 
-    const table = this._getCLITable();
+  public reportPtauFileDownloadingFinish() {
+    this._ptauFileReporter.reportDownloadingFinish();
+  }
 
-    table.push([{ content: chalk.bold("Circuit Name") }, { content: chalk.bold(`Constraints Number`) }]);
+  public reportPtauFileDownloadingError() {
+    this._ptauFileReporter.reportDownloadingError();
+  }
 
-    for (const info of compilationInfoArr) {
-      table.push([{ content: info.circuitName }, { content: info.constraintsNumber.toString(), hAlign: "right" }]);
-    }
+  public reportCircomCompilerDownloadingInfo(version: string, isWasm: boolean) {
+    this._circomCompilerReporter.reportDownloadingInfo(version, isWasm);
+  }
 
-    output += `\n${table.toString()}\n`;
+  public reportCircomCompilerDownloadingFinish() {
+    this._circomCompilerReporter.reportDownloadingFinish();
+  }
 
-    console.log(output);
+  public reportCircomCompilerDownloadingError() {
+    this._circomCompilerReporter.reportDownloadingError();
   }
 
   public reportCircuitListToSetup(
     allCircuitsSetupInfo: CircuitSetupInfo[],
     filteredCircuitsSetupInfo: CircuitSetupInfo[],
   ) {
-    if (this.isQuiet()) return;
-
-    if (filteredCircuitsSetupInfo.length > 0) {
-      let filesToSetupMessage: string = `\n${chalk.bold("Circuits to setup:")}\n`;
-
-      for (const setupInfo of filteredCircuitsSetupInfo) {
-        filesToSetupMessage += `\n${chalk.green(">")} ${chalk.italic(setupInfo.circuitArtifact.circuitSourceName)}`;
-      }
-
-      console.log(filesToSetupMessage);
-    }
-
-    const skippedFiles: CircuitSetupInfo[] = allCircuitsSetupInfo.filter(
-      (setupInfo: CircuitSetupInfo) => !filteredCircuitsSetupInfo.includes(setupInfo),
-    );
-
-    if (skippedFiles.length > 0) {
-      let skippedFilesMessage: string = `\n${chalk.bold("Setup skipped for:")}\n`;
-
-      for (const file of skippedFiles) {
-        skippedFilesMessage += `\n${chalk.yellow(">")} ${chalk.italic.grey(file.circuitArtifact.circuitSourceName)}`;
-      }
-
-      console.log(skippedFilesMessage);
-    }
+    this._setupReporter.reportCircuitList(allCircuitsSetupInfo, filteredCircuitsSetupInfo);
   }
 
   public reportSetupProcessHeader() {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.bold("Starting keys setup process:")}`;
-
-    console.log(output);
-  }
-
-  public reportPtauFileInfo(maxConstraintsNumber: number, ptauId: number, ptauFileFullPath?: string) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.italic.cyan("First step - Ptau file setup")}\n`;
-
-    output += `\nPtau file info:\n`;
-    output += `\n> Max circuits constraints number - ${maxConstraintsNumber}`;
-    output += `\n> Min required ptau ID - ${ptauId}`;
-
-    if (ptauFileFullPath) {
-      output += `\n> Found existing ptau file - ${chalk.underline(ptauFileFullPath)}`;
-    } else {
-      output += `\n> No matching ptau file was found`;
-    }
-
-    console.log(output);
-  }
-
-  public reportPtauFileDownloadingInfo(ptauFilePath: string, downloadUrl: string) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\nPtau downloading info:\n`;
-    output += `\n> Ptau file path - ${chalk.underline(ptauFilePath)}`;
-    output += `\n> Download URL - ${chalk.underline(downloadUrl)}\n`;
-
-    console.log(output);
-  }
-
-  public reportPtauFileDownloadingFinish() {
-    if (this.isQuiet()) return;
-
-    this._progressBarProcessor.stopProgressBar();
-
-    console.log(`\n${emoji("✅ ", `${chalk.green("✔ ")}`)}Ptau file successfully downloaded`);
-  }
-
-  public reportPtauFileDownloadingError() {
-    if (this.isQuiet()) return;
-
-    this._progressBarProcessor.stopProgressBar();
-
-    console.log(`\n${emoji("❌ ", `${chalk.red("X ")}`)}Ptau file downloading failed`);
-  }
-
-  public reportCircomCompilerDownloadingInfo(version: string, isWasm: boolean) {
-    if (this.isQuiet()) return;
-
-    if (isWasm) {
-      console.log(
-        `Failed to download proper platform compiler or OS platform is not supported, trying to download WASM Circom compiler v${version}`,
-      );
-    } else {
-      console.log(`No proper compiler found, trying to download the latest available Circom compiler v${version}`);
-    }
-  }
-
-  public reportCircomCompilerDownloadingFinish() {
-    if (this.isQuiet()) return;
-
-    this._progressBarProcessor.stopProgressBar();
-
-    console.log(`\n${emoji("✅ ", `${chalk.green("✔ ")}`)}Circom compiler successfully downloaded`);
-  }
-
-  public reportCircomCompilerDownloadingError() {
-    if (this.isQuiet()) return;
-
-    this._progressBarProcessor.stopProgressBar();
-
-    console.log(`\n${emoji("❌ ", `${chalk.red("X ")}`)}Circom compiler downloading failed`);
-  }
-
-  public reportZKeyFilesGenerationHeader(contributions: number) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.italic.cyan("Second step - Generating ZKey files for circuits")}\n`;
-
-    if (contributions > 0) {
-      output += `\n> Phase-2 contributions to ZKey file ${chalk.green("enabled")}`;
-      output += `\n> Contributions count: ${contributions}\n`;
-    }
-
-    output += `\nStarting generation of ZKey files:\n`;
-
-    console.log(output);
-  }
-
-  public reportZKeyFileGenerationStartWithSpinner(circuitName: string): string | null {
-    return this._startSpinner(circuitName, "generate-zkey", `Generating ZKey file for ${circuitName} circuit`);
-  }
-
-  public reportZKeyFileGenerationResult(spinnerId: string | null, circuitName: string, contributionsNumber: number) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const generationTimeMessage: string | undefined = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-    const contributionMessage: string =
-      contributionsNumber !== 0 ? `with ${contributionsNumber} ${pluralize(contributionsNumber, "contribution")} ` : "";
-
-    this._spinnerProcessor.succeedSpinner(
-      spinnerId,
-      `Generated ZKey file for ${chalk.italic(circuitName)} circuit ${contributionMessage}${generationTimeMessage}`,
-    );
-  }
-
-  public reportVKeyFilesGenerationHeader() {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.italic.cyan("Third step - Generating VKey files for circuits")}\n`;
-
-    output += `\nStarting generation of VKey files:\n`;
-
-    console.log(output);
-  }
-
-  public reportVKeyFileGenerationStartWithSpinner(circuitName: string): string | null {
-    return this._startSpinner(circuitName, "generate-vkey", `Generating VKey file for ${circuitName} circuit`);
-  }
-
-  public reportVKeyFileGenerationResult(spinnerId: string | null, circuitName: string) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const generationTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.succeedSpinner(
-      spinnerId,
-      `Generated VKey file for ${chalk.italic(circuitName)} circuit ${generationTimeMessage}`,
-    );
+    this._setupReporter.reportHeader();
   }
 
   public reportSetupResult(circuitArtifacts: CircuitArtifact[]) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-    const circuitsMessage: string = circuitArtifacts.length > 1 ? `${circuitArtifacts.length} circuits` : `one circuit`;
-
-    output += `\n${chalk.bold(`Successfully generated keys for ${circuitsMessage}.`)}\n`;
-
-    const table = this._getCLITable();
-
-    table.push([{ content: chalk.bold("Circuit Name") }, { content: chalk.bold(`ZKey file size (MB)`) }]);
-
-    for (const circuitArtifact of circuitArtifacts) {
-      table.push([
-        { content: circuitArtifact.circuitTemplateName },
-        { content: this._getFileSizeInMB(circuitArtifact.compilerOutputFiles.zkey?.fileSourcePath), hAlign: "right" },
-      ]);
-    }
-
-    output += `\n${table.toString()}\n`;
-
-    console.log(output);
+    this._setupReporter.reportResult(circuitArtifacts);
   }
 
-  public reportTypesGenerationHeaderWithSpinner(): string | null {
-    if (this.isQuiet()) return null;
-
-    let output: string = "";
-
-    output += `\n${chalk.italic.cyan("Fourth step - Generating types for circuits")}\n`;
-
-    console.log(output);
-
-    return this._startSpinner("AllCircuits", "generate-types", `Generating types for circuits`);
+  public reportZKeyFilesGenerationHeader(contributions: number) {
+    this._zKeyFilesGenerationReporter.reportHeader(contributions);
   }
 
-  public reportTypesGenerationResult(spinnerId: string | null) {
-    if (this.isQuiet() || !spinnerId) return;
+  public reportZKeyFileGenerationStartWithSpinner(circuitName: string): string | null {
+    return this._zKeyFilesGenerationReporter.reportStartWithSpinner(circuitName);
+  }
 
-    const generationTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
+  public reportZKeyFileGenerationResult(spinnerId: string | null, circuitName: string, contributionsNumber: number) {
+    this._zKeyFilesGenerationReporter.reportResult(spinnerId, circuitName, contributionsNumber);
+  }
 
-    this._spinnerProcessor.succeedSpinner(spinnerId, `Generated types for circuits ${generationTimeMessage}`);
+  public reportVKeyFilesGenerationHeader() {
+    this._vKeyFilesGenerationReporter.reportHeader();
+  }
+
+  public reportVKeyFileGenerationStartWithSpinner(circuitName: string): string | null {
+    return this._vKeyFilesGenerationReporter.reportStartWithSpinner(circuitName);
+  }
+
+  public reportVKeyFileGenerationResult(spinnerId: string | null, circuitName: string) {
+    this._vKeyFilesGenerationReporter.reportResult(spinnerId, circuitName);
   }
 
   public reportNothingToCompile() {
-    if (this.isQuiet()) return;
-
-    console.log(`\n${chalk.bold("Nothing to compile.")}\n`);
+    this._progressReporter.reportNothingTo("compile");
   }
 
   public reportNothingToSetup() {
-    if (this.isQuiet()) return;
-
-    console.log(`\n${chalk.bold("Nothing to setup.")}\n`);
+    this._progressReporter.reportNothingTo("setup");
   }
 
   public reportNothingToGenerate() {
-    if (this.isQuiet()) return;
-
-    console.log(`\n${chalk.bold("Nothing to generate.")}\n`);
+    this._progressReporter.reportNothingTo("generate");
   }
 
   public reportVerifiersGenerationHeader(verifiersType: VerifierLanguageType) {
-    if (this.isQuiet()) return;
-
-    let output: string = "";
-
-    output += `\n${chalk.bold(`Starting generation of ${this._getVerifierLanguageMessage(verifiersType)} verifier contracts...`)}\n`;
-
-    console.log(output);
+    this._verifiersGenerationReporter.reportHeader(verifiersType);
   }
 
   public reportVerifierGenerationStartWithSpinner(
     circuitName: string,
     verifiersType: VerifierLanguageType,
   ): string | null {
-    return this._startSpinner(
-      circuitName,
-      "verifier-generation",
-      `Generating ${this._getVerifierLanguageMessage(verifiersType)} verifier contract for ${circuitName} circuit`,
-    );
+    return this._verifiersGenerationReporter.reportStartWithSpinner(circuitName, verifiersType);
   }
 
   public reportVerifierGenerationResult(
@@ -430,120 +180,34 @@ class BaseReporter {
     circuitName: string,
     verifiersType: VerifierLanguageType,
   ) {
-    if (this.isQuiet() || !spinnerId) return;
-
-    const generationTimeMessage: string = this._getSpinnerWorkingTimeMessage(
-      this._spinnerProcessor.getWorkingTime(spinnerId),
-    );
-
-    this._spinnerProcessor.succeedSpinner(
-      spinnerId,
-      `Generated ${this._getVerifierLanguageMessage(verifiersType)} verifier contract for ${chalk.italic(circuitName)} circuit ${generationTimeMessage}`,
-    );
-  }
-
-  public reportStartFileDownloadingWithProgressBar(totalValue: number, startValue: number) {
-    if (this.isQuiet()) return;
-
-    this._progressBarProcessor.createAndStartProgressBar(
-      {
-        formatValue: this._progressBarProcessor.formatToMB,
-        format: "Downloading [{bar}] {percentage}% | {value}/{total} MB | Time elapsed: {duration}s",
-        hideCursor: true,
-      },
-      totalValue,
-      startValue,
-    );
+    this._verifiersGenerationReporter.reportResult(spinnerId, circuitName, verifiersType);
   }
 
   public reportVerifiersGenerationResult(verifiersType: VerifierLanguageType, verifiersCount: number) {
-    if (this.isQuiet()) return;
+    this._verifiersGenerationReporter.reportFinalResult(verifiersType, verifiersCount);
+  }
 
-    let output: string = "";
-
-    output += `\n${chalk.bold(`Successfully generated ${verifiersCount} ${this._getVerifierLanguageMessage(verifiersType)} verifier contracts.`)}\n`;
-
-    console.log(output);
+  public reportStartFileDownloadingWithProgressBar(totalValue: number, startValue: number) {
+    this._progressReporter.reportStartFileDownloadingWithProgressBar(totalValue, startValue);
   }
 
   public updateProgressBarValue(valueToAdd: number) {
-    this._progressBarProcessor.updateProgressBar(valueToAdd);
+    this._progressReporter.updateProgressBarValue(valueToAdd);
   }
 
   public verboseLog(namespace: string, formatterStr: string, logArgs: any[] = []) {
-    debug(`hardhat-zkit:${namespace}`)(formatterStr, ...logArgs);
-  }
-
-  public isQuiet(): boolean {
-    return this._quiet;
-  }
-
-  private _getCLITable(): CliTable3.Table {
-    return new CliTable3({
-      style: { head: [], border: [], "padding-left": 2, "padding-right": 2 },
-      chars: {
-        mid: "·",
-        "top-mid": "|",
-        "left-mid": " ·",
-        "mid-mid": "|",
-        "right-mid": "·",
-        left: " |",
-        "top-left": " ·",
-        "top-right": "·",
-        "bottom-left": " ·",
-        "bottom-right": "·",
-        middle: "·",
-        top: "-",
-        bottom: "-",
-        "bottom-mid": "|",
-      },
-    });
-  }
-
-  private _getSpinnerWorkingTimeMessage(workingTime: string | undefined): string {
-    return workingTime ? chalk.grey(`(${workingTime} s)`) : "";
-  }
-
-  private _getVerifierLanguageMessage(verifiersType: VerifierLanguageType): string {
-    switch (verifiersType) {
-      case "sol":
-        return "Solidity";
-      case "vy":
-        return "Vyper";
-      default:
-        throw new HardhatZKitError(`Invalid verifiers type - ${verifiersType}`);
-    }
-  }
-
-  private _getFileSizeInMB(filePath: string | undefined): string {
-    if (!filePath) {
-      throw new HardhatZKitError("File path is undefined. Unable to get file size.");
-    }
-
-    const fileSize: number = fs.statSync(filePath).size;
-
-    return (fileSize / BYTES_IN_MB).toFixed(3);
-  }
-
-  private _startSpinner(spinnerIdName: string, spinnerIdSuffix: string, spinnerText: string): string | null {
-    if (this.isQuiet()) return null;
-
-    const spinnerId: string = `${spinnerIdName}-${spinnerIdSuffix}`;
-
-    this._spinnerProcessor.createSpinner(spinnerId, { text: spinnerText });
-
-    return spinnerId;
+    this._progressReporter.verboseLog(namespace, formatterStr, logArgs);
   }
 }
 
-export let Reporter: BaseReporter | null = null;
+export let Reporter: ReporterFacade | null = null;
 
 export function createReporter(quiet: boolean) {
   if (Reporter) {
     return;
   }
 
-  Reporter = new BaseReporter(quiet);
+  Reporter = new ReporterFacade(quiet);
 }
 
 /**
